Add liveJS support.
This works by removing the old <script> tags and adding in new <script>
tags with the same src in their place. Not that jsdom cannot handle this
type of DOM manipulation, so test/options_test.coffee has several
failing tests. However, the behavior is correct on an actual DOM
implementation.